backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pyyaml==6.0.1
python-multipart==0.0.9
anthropic==0.7.8
python-dotenv==1.0.1
